requirements.txt
certifi==2018.11.29
chardet==3.0.4
defusedxml==0.5.0
Django==2.1.4
django-allauth==0.38.0
django-hashers-passlib==0.4
djangorestframework==3.9.0
idna==2.8
oauthlib==2.1.0
passlib==1.7.1
PyJWT==1.7.1
python-mimeparse==1.6.0
python3-openid==3.1.0
pytube==9.3.5
pytz==2018.7
requests==2.21.0
requests-oauthlib==1.0.0
RestAuth==0.6.3
RestAuthCommon==0.7.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==2.0.0
South==1.0.2
urllib3==1.24.1
